+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+
+## [0.5.0] - 2024-01-09
+
+### Added
+- Migrate BackButton to UI
+- Added molecules folder
+- Migrate Icon to UI
+
 ## [0.4.0] - 2024-01-09
 
 ### Added